+++ Changelog.md
@@ -1,2 +1,1 @@
-- Added fluids tags
-- Fixed crash with fluids trying to be removed from universal bin
+- Fixed pellet recipes giving too few pellets